feat: allow arbitrary unsigned types in ranges

+++ garble_examples/smart_cookie.garble.rs
@@ -12,7 +12,7 @@ pub fn log_interest(website_visit: WebsiteVisit, state: UserState) -> LogResult 
     if is_signature_ok(state, website_visit.key) {
         let interests = state.interests;
         let user_interest = website_visit.interest;
-        let updated_interests = 0u8..16u8.map(|i: usize| -> UserInterest {
+        let updated_interests = 0usize..16usize.map(|i: usize| -> UserInterest {
             if i == 0usize {
                 user_interest
             } else {
@@ -44,7 +44,7 @@ pub fn decide_ad(website_key: SigningKey, state: UserState) -> AdDecisionResult 
                 UserInterest::Arts => sums.update(5usize, sums[5] + 1u8),
             }
         });
-        let max_interest = 0u8..6u8.fold(
+        let max_interest = 0usize..6usize.fold(
             MaxInterest {
                 visits: 0u8,
                 index_of_variant: 0usize,
